[WIP]feat(skill-detail): 推移チャートの追加 #54
