+++ src/app/navigation/navigation.component.html
@@ -1,5 +1,5 @@
-<h2 mat-subheader>ランキング</h2>
-<mat-nav-list dense class="item-group">
+<mat-nav-list dense>
+  <h2 mat-subheader>ランキング</h2>
   <a mat-list-item class="item" routerLink="/">総合</a>
   <a mat-list-item class="item" routerLink="/">推移チャート</a>
   <a mat-list-item class="item" routerLink="/">条件別</a>
@@ -7,8 +7,8 @@
 
 <mat-divider></mat-divider>
 
-<h2 mat-subheader>参考情報</h2>
 <mat-nav-list dense>
+  <h2 mat-subheader>参考情報</h2>
   <a mat-list-item class="item" routerLink="/">スクール</a>
   <a mat-list-item class="item" routerLink="/">教材</a>
   <a mat-list-item class="item" routerLink="/">インフルエンサー</a>
