+++ src/app/navigation/navigation.component.scss
@@ -1,5 +1,3 @@
-h2 {
-  margin: 8px 16px 0px;
-  padding: 0px;
-  color: rgba(0, 0, 0, 0.54);
+mat-nav-list {
+  margin-top: 8px;
 }
