+++ src/app/shared/shared.module.ts
@@ -4,6 +4,8 @@ import { FilterFormComponent } from './filter-form/filter-form.component';
 import { BannerComponent } from './banner/banner.component';
 import { ResultToSkillPipe } from '../pipes/result-to-skill.pipe';
 import { ResultToChartDataPipe } from '../pipes/result-to-chart-data.pipe';
+import { TransitionChartComponent } from './transition-chart/transition-chart.component';
+import { NgxChartsModule } from '@swimlane/ngx-charts';
 
 @NgModule({
   declarations: [
@@ -11,13 +13,15 @@ import { ResultToChartDataPipe } from '../pipes/result-to-chart-data.pipe';
     BannerComponent,
     ResultToSkillPipe,
     ResultToChartDataPipe,
+    TransitionChartComponent,
   ],
-  imports: [CommonModule],
+  imports: [CommonModule, NgxChartsModule],
   exports: [
     FilterFormComponent,
     BannerComponent,
     ResultToSkillPipe,
     ResultToChartDataPipe,
+    TransitionChartComponent,
   ],
 })
 export class SharedModule {}
