+++ src/app/shared/transition-chart/transition-chart.component.html
@@ -1,3 +1,5 @@
+<p>transition-chart works!</p>
+
 <ngx-charts-line-chart
   [view]="view"
   [scheme]="colorScheme"
@@ -9,9 +11,11 @@
   [xAxisLabel]="xAxisLabel"
   [yAxisLabel]="yAxisLabel"
   [timeline]="timeline"
-  [results]="chartData"
+  [results]="multi"
   (select)="onSelect($event)"
   (activate)="onActivate($event)"
   (deactivate)="onDeactivate($event)"
 >
 </ngx-charts-line-chart>
+
+<p>transition-chart end.</p>
