+++ src/app/skill-detail/skill-detail-transition/skill-detail-transition.component.html
@@ -1,6 +1,4 @@
 <p>skill-detail-transition works!</p>
 <p>skillCaption:{{ skill.skillCaption }}</p>
 
-<!-- <app-transition-chart
-  [chartData]="transitionResults | resultToChartData"
-></app-transition-chart> -->
+<app-transition-chart></app-transition-chart>
