+++ src/app/skill-detail/skill-detail.module.ts
@@ -12,11 +12,6 @@ import { MatTabsModule } from '@angular/material/tabs';
 import { SkillDetailDistributionComponent } from './skill-detail-distribution/skill-detail-distribution.component';
 import { SkillDetailBreakdownComponent } from './skill-detail-breakdown/skill-detail-breakdown.component';
 import { SkillDetailTransitionComponent } from './skill-detail-transition/skill-detail-transition.component';
-// import { BrowserModule } from '@angular/platform-browser';
-// import { FormsModule } from '@angular/forms';
-// import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { NgxChartsModule } from '@swimlane/ngx-charts';
-import { TransitionChartComponent } from './skill-detail-transition/transition-chart/transition-chart.component';
 
 @NgModule({
   declarations: [
@@ -24,7 +19,6 @@ import { TransitionChartComponent } from './skill-detail-transition/transition-c
     SkillDetailDistributionComponent,
     SkillDetailBreakdownComponent,
     SkillDetailTransitionComponent,
-    TransitionChartComponent,
   ],
   imports: [
     CommonModule,
@@ -34,10 +28,6 @@ import { TransitionChartComponent } from './skill-detail-transition/transition-c
     MatButtonModule,
     MatIconModule,
     MatTabsModule,
-    // BrowserModule,
-    // FormsModule,
-    // BrowserAnimationsModule,
-    NgxChartsModule,
   ],
 })
 export class SkillDetailModule {}
